+++ apps/maimai-web/src/lib/components/profile/Rating.svelte
@@ -9,27 +9,29 @@
   let { rating, calculatedRating }: Props = $props();
 
   let ratingLevel = $derived(
-    rating >= 15000
-      ? "rainbow"
-      : rating >= 14500
-        ? "platinum"
-        : rating >= 14000
-          ? "gold"
-          : rating >= 13000
-            ? "silver"
-            : rating >= 12000
-              ? "bronze"
-              : rating >= 10000
-                ? "purple"
-                : rating >= 7000
-                  ? "red"
-                  : rating >= 4000
-                    ? "orange"
-                    : rating >= 2000
-                      ? "green"
-                      : rating >= 1000
-                        ? "blue"
-                        : "normal",
+    rating >= 16000
+      ? "rainbow_kiwami"
+      : rating >= 15000
+        ? "rainbow"
+        : rating >= 14500
+          ? "platinum"
+          : rating >= 14000
+            ? "gold"
+            : rating >= 13000
+              ? "silver"
+              : rating >= 12000
+                ? "bronze"
+                : rating >= 10000
+                  ? "purple"
+                  : rating >= 7000
+                    ? "red"
+                    : rating >= 4000
+                      ? "orange"
+                      : rating >= 2000
+                        ? "green"
+                        : rating >= 1000
+                          ? "blue"
+                          : "normal",
   );
 
   let ratingMatched = $derived(rating === calculatedRating);