+++ apps/maimai-web/src/lib/constants/pureMapping.ts
@@ -1,9 +1,12 @@
 export const backgroundMapping = {
   CiRCLE: "/assets/circle/bg.webp",
+  "CiRCLE+": "/assets/circle+/bg.webp",
 };
 export const logoMapping = {
   CiRCLE: "/assets/circle/logo.webp",
+  "CiRCLE+": "/assets/circle+/logo.webp",
 };
 export const versionNameMapping = {
   CiRCLE: "CiRCLE",
+  "CiRCLE+": "CiRCLE+",
 };